HEEDLS-492 Refactor AllDelegatesViewModel Filters tests
