+++ DigitalLearningSolutions.Web.Tests/ViewModels/TrackingSystem/Delegates/AllDelegates/AllDelegatesViewModelTests.cs
@@ -5,22 +5,14 @@
     using System.Linq;
     using DigitalLearningSolutions.Data.Models.CustomPrompts;
     using DigitalLearningSolutions.Data.Models.User;
-    using DigitalLearningSolutions.Data.Services;
-    using DigitalLearningSolutions.Data.Tests.TestHelpers;
     using DigitalLearningSolutions.Web.Helpers;
-    using DigitalLearningSolutions.Web.Helpers.FilterOptions;
-    using DigitalLearningSolutions.Web.Models.Enums;
     using DigitalLearningSolutions.Web.ViewModels.Common.SearchablePage;
     using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.AllDelegates;
-    using FakeItEasy;
     using FluentAssertions;
     using NUnit.Framework;
 
     public class AllDelegatesViewModelTests
     {
-        private readonly CentreCustomPromptHelper customPromptHelper;
-        private readonly ICentreCustomPromptsService customPromptsService;
-
         private readonly DelegateUserCard[] delegateUsers =
         {
             new DelegateUserCard
@@ -54,12 +46,6 @@
             new DelegateUserCard
                 { FirstName = "o", LastName = "Surname", DateRegistered = DateTime.Today.AddDays(1) }
         };
-
-        public AllDelegatesViewModelTests()
-        {
-            customPromptsService = A.Fake<ICentreCustomPromptsService>();
-            customPromptHelper = new CentreCustomPromptHelper(customPromptsService);
-        }
 
         [Test]
         public void All_delegates_should_default_to_returning_the_first_ten_delegates()
@@ -96,141 +82,6 @@
 
             model.Delegates.Count().Should().Be(5);
             model.Delegates.First().DelegateInfo.Name.Should().BeEquivalentTo("k Surname");
-        }
-
-        [Test]
-        public void All_delegates_filters_should_be_set()
-        {
-            // Given
-            var passwordStatusOptions = new[]
-            {
-                DelegatePasswordStatusFilterOptions.PasswordSet,
-                DelegatePasswordStatusFilterOptions.PasswordNotSet
-            };
-
-            var adminStatusOptions = new[]
-            {
-                DelegateAdminStatusFilterOptions.IsAdmin,
-                DelegateAdminStatusFilterOptions.IsNotAdmin
-            };
-
-            var activeStatusOptions = new[]
-            {
-                DelegateActiveStatusFilterOptions.IsActive,
-                DelegateActiveStatusFilterOptions.IsNotActive
-            };
-
-            var jobGroups = new List<(int id, string name)>
-                { (1, "J 1"), (2, "J 2") };
-            var jobGroupOptions = new[]
-            {
-                new FilterOptionViewModel(
-                    "J 1",
-                    "JobGroupId" + FilteringHelper.Separator +
-                    "JobGroupId" + FilteringHelper.Separator + 1,
-                    FilterStatus.Default
-                ),
-                new FilterOptionViewModel(
-                    "J 2",
-                    "JobGroupId" + FilteringHelper.Separator +
-                    "JobGroupId" + FilteringHelper.Separator + 2,
-                    FilterStatus.Default
-                )
-            };
-
-            var registrationTypeOptions = new[]
-            {
-                DelegateRegistrationTypeFilterOptions.SelfRegistered,
-                DelegateRegistrationTypeFilterOptions.SelfRegisteredExternal,
-                DelegateRegistrationTypeFilterOptions.RegisteredByCentre,
-            };
-
-            var customPrompt1 = CustomPromptsTestHelper.GetDefaultCustomPrompt(
-                1,
-                "First prompt",
-                "Clinical\r\nNon-Clinical"
-            );
-            var customPrompt3 = CustomPromptsTestHelper.GetDefaultCustomPrompt(3);
-            var customPrompt4 = CustomPromptsTestHelper.GetDefaultCustomPrompt(4, "Fourth prompt", "C 1\r\nC 2\r\nC 3");
-            var centreCustomPrompts = new CentreCustomPrompts(
-                1,
-                new List<CustomPrompt> { customPrompt1, customPrompt3, customPrompt4 }
-            );
-
-            var customPrompt1Options = new[]
-            {
-                new FilterOptionViewModel(
-                    "Clinical",
-                    "Answer1" + FilteringHelper.Separator +
-                    "Answer1" + FilteringHelper.Separator + "Clinical",
-                    FilterStatus.Default
-                ),
-                new FilterOptionViewModel(
-                    "Non-Clinical",
-                    "Answer1" + FilteringHelper.Separator +
-                    "Answer1" + FilteringHelper.Separator + "Non-Clinical",
-                    FilterStatus.Default
-                ),
-                new FilterOptionViewModel(
-                    "No option selected",
-                    "Answer1" + FilteringHelper.Separator +
-                    "Answer1" + FilteringHelper.Separator + FilteringHelper.EmptyValue,
-                    FilterStatus.Default
-                )
-            };
-            var customPrompt4Options = new[]
-            {
-                new FilterOptionViewModel(
-                    "C 1",
-                    "Answer4" + FilteringHelper.Separator +
-                    "Answer4" + FilteringHelper.Separator + "C 1",
-                    FilterStatus.Default
-                ),
-                new FilterOptionViewModel(
-                    "C 2",
-                    "Answer4" + FilteringHelper.Separator +
-                    "Answer4" + FilteringHelper.Separator + "C 2",
-                    FilterStatus.Default
-                ),
-                new FilterOptionViewModel(
-                    "C 3",
-                    "Answer4" + FilteringHelper.Separator +
-                    "Answer4" + FilteringHelper.Separator + "C 3",
-                    FilterStatus.Default
-                ),
-                new FilterOptionViewModel(
-                    "No option selected",
-                    "Answer4" + FilteringHelper.Separator +
-                    "Answer4" + FilteringHelper.Separator + FilteringHelper.EmptyValue,
-                    FilterStatus.Default
-                )
-            };
-
-            var expectedFilters = new[]
-            {
-                new FilterViewModel("PasswordStatus", "Password Status", passwordStatusOptions),
-                new FilterViewModel("AdminStatus", "Admin Status", adminStatusOptions),
-                new FilterViewModel("ActiveStatus", "Active Status", activeStatusOptions),
-                new FilterViewModel("JobGroupId", "Job Group", jobGroupOptions),
-                new FilterViewModel("RegistrationType", "Registration Type", registrationTypeOptions),
-                new FilterViewModel("CustomPrompt1", "First prompt", customPrompt1Options),
-                new FilterViewModel("CustomPrompt4", "Fourth prompt", customPrompt4Options)
-            }.AsEnumerable();
-
-            // When
-            var model = new AllDelegatesViewModel(
-                delegateUsers,
-                jobGroups,
-                centreCustomPrompts.CustomPrompts,
-                2,
-                null,
-                DelegateSortByOptions.Name.PropertyName,
-                BaseSearchablePageViewModel.Ascending,
-                null
-            );
-
-            // Then
-            model.Filters.Should().BeEquivalentTo(expectedFilters);
         }
 
         [Test]
